General tidy-up (large commit)
This commit makes a number of small changes to various aspects of the
Specifications. These changes are the result of some changes to
underlying data model and the tools used to process that data. This will
make it easier to write and read Specifications in the future.

The most noticeable change is that the HTML view has been re-formatted.
However, the content is generally unchanged and its meaning is
unaltered.

The tables of Terms and External References towards the bottom of the
HTML page has been redesigned to make it easier to see which documents
define which Terms.

The `id`s of nodes in the Specification tree in the HTML page will now
have IDs that are more stable. This will minimise differences in the
HTML page between git commits.

Some non-ASCII characters were being stripped out of the old rendering,
even though the page was UFT-8 encoded. These characters are now
included properly.

Files distributed with the Specification package are now listed in the
correct order (for example, the order in which they are imported by a
"parent" XML Schema).